+++ src/tests/picker-keys.test.ts
@@ -2,12 +2,13 @@ import { expect, test } from "bun:test";
 import { parseKeys } from "../picker-keys";
 
 test("parses picker navigation keys", () => {
-  expect(parseKeys("\x1b[A\x1b[B\x0e\x0a\x10\t\r\x7f")).toEqual([
+  expect(parseKeys("\x1b[A\x1b[B\x0e\x0a\x10\x19\t\r\x7f")).toEqual([
     "up",
     "down",
     "ctrl-n",
     "ctrl-j",
     "ctrl-p",
+    "ctrl-y",
     "tab",
     "enter",
     "backspace",